backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.44
psycopg[binary]==3.2.13
pydantic-settings==2.7.0
alembic==1.14.0
email-validator==2.2.0
firebase-admin==6.6.0
httpx==0.28.1
